Add guild disbanding Add better logging to guild functions Fix bug with GetGuildByID function Refactor object iteration to use dictionary lookup base on PR feedback Split most guild handlers into new file Add better handling of missing guilds in message responses Add handler for OperateGuild messages
